code reorganizations
