+++ VenstarTranslator/Models/SensorPurpose.cs
@@ -1,0 +1,9 @@
+namespace VenstarTranslator.Models;
+
+public enum SensorPurpose
+{
+    Outdoor = 1,
+    Return = 2,
+    Remote = 3,
+    Supply = 4,
+}
