+++ VenstarTranslator/Models/TemperatureScale.cs
@@ -1,0 +1,7 @@
+namespace VenstarTranslator.Models;
+
+public enum TemperatureScale
+{
+    F = 1,
+    C = 2
+}
