+++ VenstarTranslator/Models/TranslatedVenstarSensor.cs
@@ -9,13 +9,12 @@ using System.Text.RegularExpressions;
 
 using Hangfire;
 
-using Microsoft.EntityFrameworkCore;
-
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Linq;
 
 using VenstarTranslator.Models.Protobuf;
+using VenstarTranslator.Models.Validation;
 
 namespace VenstarTranslator.Models;
 
@@ -187,34 +186,4 @@ public class TranslatedVenstarSensor
         }
         return Convert.ToDouble(target);
     }
-}
-
-[Owned]
-public class DataSourceHttpHeader
-{
-    [JsonIgnore]
-    [Key]
-    public int ID { get; set; }
-
-    [JsonProperty(Order = 1)]
-    [Required(AllowEmptyStrings = false, ErrorMessage = "Header name is required.")]
-    public string Name { get; set; }
-
-    [JsonProperty(Order = 2)]
-    [Required(AllowEmptyStrings = false, ErrorMessage = "Header value is required.")]
-    public string Value { get; set; }
-}
-
-public enum SensorPurpose
-{
-    Outdoor = 1,
-    Return = 2,
-    Remote = 3,
-    Supply = 4,
-}
-
-public enum TemperatureScale
-{
-    F = 1,
-    C = 2
 }
